pyproject.toml
[tool.poetry]
name = "blondie"
version = "0.1.0"
description = "Autonomous coding agent"
authors = ["iva2k <iva2k@yahoo.com>"]
readme = "README.md"
maintainers = ["iva2k <iva2k@yahoo.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^2.8.0"
pyyaml = "^6.0.2"
types-PyYAML = "^6.0.12.20250915"
gitpython = "^3.1.43"
openai = "^1.51.0"
anthropic = "^0.34.2"
httpx = "^0.27.0"
click = "^8.1.7"
rich = "^13.9.3"
pathspec = "^0.12.1"
markdown = "^3.7"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
black = "^24.8.0"
ruff = "^0.6.9"
mypy = "^1.11.2"
pylint = "^3.2.7"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = false
warn_unused_configs = true
disallow_untyped_defs = false
install_types = true
